requirements.txt
fastapi==0.115.9
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
python-telegram-bot==21.10
